Auto-PR: Consolidate duplicate formatDuration into utils/formatters.ts (closes #501)
